requirements.txt
Pillow~=8.2.0
requests~=2.25.1
selenium~=3.141.0
websockets~=9.0.2
baidu-aip==2.2.18.0
numpy~=1.20.3
easyocr==1.3.2
PyYAML~=5.4.1
urllib3~=1.26.5